+++ R/master_spin.R
@@ -26,69 +26,110 @@ master_spin <-
            stop = Inf,
            logdir = config::get("dirs")$logs,
            format = "html") {
-
     scriptlist = dir(".", pattern = pattern, full.names = T)
     tostop = min(length(scriptlist), stop)
+
     reslogs <-
       lapply(scriptlist[start:tostop], function(x) {
         cat(x, "\n")
+
+        # Define if a signature and session infor should be included
+        need_a_signature <- TRUE
+        need_a_sessioninfo <- TRUE
+        rlx <- readLines(x)
+        headlin <- grep("\\#' ---", rlx)
+        # No well format heading
+        if (length(headlin) != 2) {
+          need_a_signature = FALSE
+          nedd_a_sessioninfo = FALSE
+        }
+        else {
+          yamlx <- yaml.load(
+            gsub("\\#' ", "", rlx[seq(headlin[1] + 1, headlin[2] - 1)]))
+          if (exists("yamalx$signature")) {
+            need_a_signature = identical(yamalx$signature, TRUE)
+          }
+          else {
+            need_a_signature = FALSE
+          }
+
+          if (exists("yamalx$sessioninfo")) {
+            need_a_sessioninfo = identical(yamalx$sessioninfo, TRUE)
+          }
+          else {
+            need_a_sessioninfo = FALSE
+          }
+        }
+
+
+        # Add a Session info is required
+        if (need_a_sessioninfo) {
+          rlx[length(rlx) + 1] <- paste0("#' ## Session Info ",
+                                         "\n",
+                                         "#+ echo = F ",
+                                         "\n",
+                                         "print(sessionInfo(), locale = F)")
+        }
+
+
+        # Add signature if required
+        if (need_a_signature) {
+          hash <- digest(x, algo = "sha1", file = T)
+          start_var <-
+            file_path_sans_ext(basename(tempfile("time")))
+          rlx[headlin[2] + 1] <- paste0("#+ echo = F\n",
+                                        start_var, "<- Sys.time()\n",
+                                        rlx[headlin[2] + 1])
+          rlx[length(rlx) + 1] <- paste0(
+            "#' ## Signature ",
+            "\n",
+            "#+ echo = F ",
+            "\n",
+            "cat(",
+            "\n",
+            "'File Name: ",
+            x,
+            "\\n'," ,
+            "\n",
+            "'SHA1: ",
+            hash,
+            "\\n',",
+            "\n",
+            "'Execution start: ', format(",
+            start_var,
+            "), '\\n',",
+            "\n",
+            "'Execution time: ', round(difftime(Sys.time(),",
+            start_var,
+            ", units = 'min'),2), 'min')"
+          )
+        }
+
+        fnx <- paste0(basename(tempfile()), ".R")
+        bnx <- file_path_sans_ext(basename(x))
+        writeLines(rlx, fnx)
         start = Sys.time()
         res <-
-          try(render_report(x,format,logdir))
-        res <- ifelse(is.null(res),0,res)
+          try(
+            render_report(
+              fnx,
+              format,
+              logdir,
+              output_file = file.path(logdir, bnx)))
+        file.remove(fnx)
+        res <- ifelse(is.null(res), 0, res)
         end <- Sys.time()
         elapsed = difftime(end, start)
 
-        # Define if a signature should be included
-        need_a_signature <- TRUE
-        rlx <- readLines(x)
-        headlin <- grep("\\#' ---",rlx)
-        # No well format heading
-        if (length(headlin) != 2) {
-            need_a_signature = FALSE
-        }
-        else {
-         yamlx <-yaml.load(
-           gsub("\\#' ","", readLines(x)[seq(headlin[1]+1,headlin[2]-1)]))
-         if (is.null(yamlx$signature) | ! identical(yamlx$signature,TRUE))
-             need_a_signature = FALSE
-        }
-        if (need_a_signature) {
-          # Add the signature
-          bsname <- basename(file_path_sans_ext(x))
-          lsname <- file.path(logdir,paste0(bsname,".html"))
-          hash <- digest(x, algo = "sha1", file = T)
-          if (format == "html" & file.exists(lsname)) {
-            pattern <- "(.*)(</div>)"
-            frl <- readLines(lsname)
-            sig <- paste0(
-              "<p></p>\n",
-              "<pre><code>",
-              paste0(rep("=", 80), collapse = ""),"\n",
-              "FILENAME:        ",x, "\n",
-              "SHA1:          ", hash, "\n",
-              "START TIMESTAMP: ", format(start) , "\n" ,
-              "END TIMESTAMP:   ", format(end), "\n",
-              paste0(rep("=", 80), collapse = ""),
-              "</code></pre>\n",
-              "</div>"
-            )
-            lwt <- max(grep(pattern, frl))
-            frl[lwt] <- gsub(pattern,sig,frl[lwt], perl = TRUE)
-            writeLines(frl, con = lsname)
-          }
-        }
         print(format(start))
         print(x)
         print(format(elapsed))
-        print(ifelse(inherits(res, "try-error") |
-                       res != 0, "FAIL", ":-)"))
+        print(ifelse(inherits(res, "try-error") | res != 0, "FAIL", ":-)"))
         data.frame(
           timestart = format(start),
           script = x,
           elapsed = format(elapsed),
-          comments = ifelse(inherits(res, "try-error") |
-                              res != 0, "FAIL", ":-)")
+          comments = ifelse(inherits(res, "try-error") | res != 0, "FAIL", ":-)")
         )
       })