+++ tests/testthat/test_master_spin.R
@@ -21,7 +21,7 @@ test_that("The pipeline works fine", {
   file.copy(from = dir(curdir, pattern = "^[0-9][0-9].*\\.R$", full.names = T), to = tmpfile)
 
   dfres <- master_spin(format = "pdf")
-  expect_true(file.exists("logs/01_clean.pdf"))
+  expect_true(file.exists("logs/00_clean.pdf"))
   expect_true(file.exists("logs/02_prog_a.pdf"))
   expect_true(file.exists("logs/03_prog_b.pdf"))
   expect_false(file.exists("logs/04_prog_c.pdf"))
@@ -52,7 +52,7 @@ test_that("The pipeline works fine with html", {
   file.copy(from = dir(curdir, pattern = "^[0-9][0-9].*\\.R$", full.names = T), to = tmpfile)
 
   dfres <- master_spin(format = "html")
-  expect_true(file.exists("logs/01_clean.html"))
+  expect_true(file.exists("logs/00_clean.html"))
   expect_true(file.exists("logs/02_prog_a.html"))
   expect_true(file.exists("logs/03_prog_b.html"))
   expect_false(file.exists("logs/04_prog_c.html"))